requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-swagger-ui==4.11.1
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
pymongo==4.5.0
weaviate-client==3.25.0
langchain==0.0.335
langchain_openai==0.0.2
requests==2.31.0
numpy==1.25.2
pandas==2.1.0
tqdm==4.66.1
mistralai==0.0.7
pytest==7.4.2
gunicorn==21.2.0
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==10.1.0 